requirements.txt
Flask==2.0.3
psycopg2==2.9.3
gunicorn==20.1.0
SQLAlchemy==1.4.31
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-CORS==3.0.10
PyJWT==2.3.0
flask-restx==0.5.1